requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attr==0.3.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
blinker==1.4
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
Click==7.0
colorama==0.4.3
coremltools==4.1
cryptography==2.8
cycler==0.10.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mpmath==1.2.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pyspark==3.1.2
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
sympy==1.8
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.61.1
typing-extensions==3.7.4.3
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.4.2
